Inwards F1
Progress:
There is now a How To screen and images showing the enemies.
There is now an end game phase including the end story and a thanks for playing screen. 
I've updated the sprites so the enemies and kofu all look different. The tilemap is also improved, each level has a different flooring.
I've renamed files and cleaned them up as well for the submission on the 4th of Nov Thurs.
I changed the font to match the game better.
All the screens now all have a different background that makes the game look much nicer.
The buttons now can be controlled by keyboard instead of clicking for ease of access. Did this by adding a focus grab on every button on each screen.
There is now a hit effect when the player or enemy get hit.
The levels are all completed with the new enemies to go with.
The sound effects are improved and fixed the sounds that were overpowering the quieter sounds.
I added a way for the player to get to the next level by standing on the bed and pressing 'K'.